refactor to have src dir
