+++ .nvmrc
@@ -1,1 +1,1 @@
-22
+24
