+++ package.json
@@ -70,7 +70,7 @@
     "typescript-eslint": "^8.57.2"
   },
   "engines": {
-    "node": ">=22",
+    "node": ">=24",
     "pnpm": ">=10.0.0"
   },
   "funding": {
