+++ src/app/sitemap.ts
@@ -1,6 +1,6 @@
 export const dynamic = 'force-static';
 
-import { clientEnv } from '@/clientEnv';
+import { clientEnv } from '@/src/clientEnv';
 import { getPublishedPosts } from '@/src/lib/posts';
 import { MetadataRoute } from 'next';
