+++ src/content/blog/building-a-chrome-extension-with-shared-state.mdx
