+++ src/content/documentation/keyboard-shortcut-setup.mdx
