+++ src/lib/posts.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import matter from 'gray-matter';
 import readingTime from 'reading-time';
-import { clientEnv } from '@/clientEnv';
+import { clientEnv } from '@/src/clientEnv';
 
 export interface Post {
   title: string;
